fix: Use asyncExecute for OBKV queries and convert Decimal to Double
Three fixes for OBKV integration test failures:

1. Use asyncExecute() instead of execute() for OBKV table queries.
   The OBKV client's simple query mode does not support BatchSize;
   asyncExecute() enables streaming batch result fetching.

2. Convert DecimalType to Double for OBKV write operations.
   The obkv-table-client-java does not support BigDecimal/DECIMAL type
   (ObNumberType is marked TODO with empty encode/decode). DECIMAL
   columns in test tables are changed to DOUBLE as a workaround.

3. Handle BigDecimal conversion in V1 non-catalog write path to avoid
   IllegalArgumentException from ObObjType.valueOfType().

+++ spark-connector-oceanbase/spark-connector-oceanbase-base/src/test/resources/sql/mysql/obkv_test.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS obkv_products (
   id          INTEGER      NOT NULL,
   name        VARCHAR(255) NOT NULL DEFAULT 'default',
   description VARCHAR(512),
-  weight      DECIMAL(20, 10),
+  weight      DOUBLE,
   PRIMARY KEY(id)
 );
 
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS obkv_all_types (
   col_bigint  BIGINT,
   col_float   FLOAT,
   col_double  DOUBLE,
-  col_decimal DECIMAL(20, 5),
+  col_decimal DOUBLE,
   col_varchar VARCHAR(255),
   col_date    DATE,
   col_ts      TIMESTAMP    NULL DEFAULT NULL
